refactor: 폴더 구조
