--- src/pages/admin/media/Media.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import AdminLayout from 'src/layouts/admin/main/AdminLayout';
-
-const AdminMedia = () => {
-  return (
-    <AdminLayout>
-      <div>관리자 페이지에 오신 것을 환영합니다.</div>
-    </AdminLayout>
-  );
-};
-
-export default AdminMedia;
