+++ src/pages/admin/media/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const AdminMedia = () => {
+  return <div>관리자 페이지에 오신 것을 환영합니다.</div>;
+};
+
+export default AdminMedia;
